requirements.txt
torch
transformers
Pillow
opencv-python-headless
ultralytics
pyyaml
numpy
tqdm
google-generative-ai

